sys_power: Ignore individual EC error response firmware errors.
These errors are only meaningful when they result in a higher-level
error. In cases where the caller does not consider it an error for the
EC to return an error for a particular EC request, such as reading a CBI
field and gracefully falling back to a default value when unset,
treating these log messages as an error results in spurious test
failures. If the firmware considers this EC response indicative of a
real error, it is responsible for logging a meaningful error to be
captured as a FirmwareError in tests.

BUG=b:233972322
TEST=power_Resume on nissa models

Change-Id: I240274f7ce084a77d2335b27fc87772d65a07501
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3765402
Tested-by: Sam McNally <sammc@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Reviewed-by: Reka Norman <rekanorman@chromium.org>
Reviewed-by: Peter Marheine <pmarheine@chromium.org>
Commit-Queue: Sam McNally <sammc@chromium.org>