connectivity_common: support the version_check option for test_that
This patch supports a new version_check option for test_that. If the
option is set to false as

  test_that --args "version_control=false" ...

it would not check the autotest commit against current branch. The
autotest could be executed on any local autotest branch even the
branch commit does not match the ToT's or the even the branch is not
clean. This would be handy for a test developer to run a local
working autotest branch without the hack of commenting out
test_version_setup() in the test control files.

BUG=b:239891832
TEST=In a local autotest branch, run the AVL tests
(cr) test_that --autotest_dir ~/trunk/src/third_party/autotest/files/
     --args "version_check=false" $DUT_IP bluetooth_AdapterQuickHealth.AVL

Change-Id: Idc8d9e55c3efad35cc5f9632e8607bab8b350643
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3782024
Reviewed-by: Shijin Abraham <shijinabraham@google.com>
Reviewed-by: Archie Pusaka <apusaka@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>