tradefed_test: Add JAVA_TOOL_OPTIONS=-XX:MaxRAM=40g when launching tradefed in lab
The Java process is getting SIGKILLed likely due to OOM when it is run
in the lab. Put a cap with the -XX:MaxRAM=40g option.
Since cts-tradefed does not provide an option to inject Java options,
JAVA_TOOL_OPTIONS is set as an environment variable.

BUG=b:237990639
TEST=utils/unittest_suite.py  server.cros.tradefed.tradefed_test_unittest

Change-Id: I096bd4cc24c32d67d00ac9c37e7b3cf3b84efae5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3782026
Reviewed-by: Kazuhiro Inaba <kinaba@chromium.org>
Reviewed-by: Shao-Chuan Lee <shaochuan@chromium.org>
Auto-Submit: Rintaro Kuroiwa <rkuroiwa@google.com>
Tested-by: Rintaro Kuroiwa <rkuroiwa@google.com>