autotest: remove BluetoothScanPerf.*_AVL from bluetooth_wifi_coex suite
Scheduled infra runners do not have a git sourcetree locally. These AVL
tests requires git on the test runner in order to move the checkout to
the right commit. As a result, these control files cannot run on infra.
Removing the suite labeler such that when the suite is scheduled, it
does not run these subtests.

These not only can't run on infra, they also shouldn't run on infra since they're functionally identical to the non _AVL version. The AVL mechanism is meant to ensure partners run the tests with a clean suite.

BUG=b:239503732
TEST=None (please let me know how I can test this)

Change-Id: I9bd2feba481316c1e377bffbd60b518ca1c68cf1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3783194
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Reviewed-by: Kevin Lund <kglund@google.com>
Commit-Queue: Billy Zhao <billyzhao@chromium.org>
Tested-by: Billy Zhao <billyzhao@chromium.org>