bluetooth: add new sr_suspend_delay_while_receiving_bqr test
A new suspend test is added to enforce a suspend delay in the
pre-suspend phase. The test connects to a classic keyboard device
and an audio device before suspend and make sure that BQR events
do not prevent the system from suspending and resuming correctly.

BUG=b:237256795
TEST=Run the test
(cr) test_that --args "btpeer_host_list=${BTPEER1},${BTPEER2}" ${DUT}
     bluetooth_AdapterSRHealth.sr_suspend_delay_while_receiving_bqr

- On a dewatt with guybrush-release/R106-15001.0.0,
  the above test passed.
- After emerging and deploying new power_manager by reverting patch
      power: Handle Bluetooth quirk with autosuspend
  the above test failed with the powerd message

  INFO powerd: [wakeup_device.cc(70)] Device
       /sys/devices/pci0000:00/0000:00:08.1/0000:03:00.4/usb3/3-3 had
       event_count 2 before suspend and 4 after resume
  INFO powerd: [dark_resume.cc(40)] User triggered wake

Hence, this test is effective in capturing the suspend regression
described in the bug b:237256795.

Change-Id: I3d25819c85abf863db024151b65e5e8ef239cff6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3787939
Reviewed-by: Ying Hsu <yinghsu@chromium.org>
Reviewed-by: Archie Pusaka <apusaka@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>