Add new scroll latency metrics to power_test
Older Event.Latency.* metrics are being deprecated in favor of newer
EventLatency.* metrics. The main difference is that the old metrics
measure up to when the frame caused by the event is swapped to GPU while
the new metrics measure up to when the swapped frame is actually
presented to the user. This CL adds new
EventLatency.KeyPressed.TotalLatency metric to power_test alongside the
old one. After a few weeks, when have some data to compare the two
metrics and deduce new thresholds for the new metric, we can remove the
old one.

BUG=chromium:1054021
TEST=Ran the test using crosfleet

Change-Id: Ib4e5915113e84fa662add006afafbf39e5315de3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3779642
Tested-by: Mohsen Izadi <mohsen@chromium.org>
Commit-Queue: Mohsen Izadi <mohsen@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>