bluetooth: add new sr_suspend_delay test
A new suspend test is added to enforce a suspend delay in the
pre-suspend phase. The test connects to an BLE mouse device
and an audio device before suspend and make sure that HCI events
do not prevent the system from suspending and resuming correctly.

In this test, the BQR feature is not enabled. Hence, the test is
expected to run on all models.

BUG=b:237256795
TEST=Run the test
(cr) test_that --args "btpeer_host_list=${BTPEER1},${BTPEER2}" ${DUT}
     bluetooth_AdapterSRHealth.sr_suspend_delay

Change-Id: Ie283ececc35a7fe3518476f3d04e6cdfd712a50d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3793527
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Ying Hsu <yinghsu@chromium.org>
Reviewed-by: Archie Pusaka <apusaka@chromium.org>