OOBE: Remove the waiting for EULA next button to be enabled
EULA next button is always enabled after: crrev.com/c/3788198

Although the user was never able to see the next button in a disabled
state before the change, the next button was only enabled when the
screen advances to the `loaded` step.
So, waiting for the button to be enabled ensured that we're past the
loading screen and the loaded step is shown.

To wait for the loaded step to be shown, use
EulaScreen.isReadyForTesting() instead of waiting for
EulaScreen.isVisible()

Bug: b:241104698
Change-Id: I66796c99aadf515e739ba02345ba975f0a2f1205
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3806867
Reviewed-by: Roman Sorokin <rsorokin@google.com>
Commit-Queue: Osama Fathy <osamafathy@google.com>
Tested-by: Osama Fathy <osamafathy@google.com>