hwqual: use 90% constraint for charge amount
The power state for some models will report that the device battery is
"Fully Charged" before the percentage charge reaches the 100% level. At
this point, the battery charge test will exit. For these devices, there
may be a final charge percentage of 97%, and a start charge of 4-5%,
which fails the current 94% charge requirement. Change the threshold to
90% to ensure these devices pass.

BUG=b:239537499
TEST=run test locally with test_that

Change-Id: I5b1f23b99b9f9625b4b12e94e328b720ff4cd3f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3840709
Tested-by: Parker Holloway <jparkerh@google.com>
Commit-Queue: Parker Holloway <jparkerh@google.com>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Reviewed-by: Chen-Tsung Hsieh <chentsung@chromium.org>