autotest: Update expected_performance_results for StP2 devices
The expected_performance_results for devices with StP2 WiFi are out of
date.

The expected values for UDP_RX on Eve and Nami should both be ~200 mbps.
Crosbolt source:
http://croshealthmon/time_series/network_WiFi_Perf.vht80_pmf/ch044_modeVHT80_psk_PSon_governor-default.udp_rx?suite=wifi&skuFilter=caroline_intel_skylake_core_m3_4GB%7Cakali_IntelR_CoreTM_i5_8250U_CPU_1_60GHz_8GB%7Ceve_IntelR_CoreTM_i7_7Y75_CPU_1_30GHz_16GB

The comments for these expectations also point to the wrong bug. Update
them to point to the bug that is actually tracking fixing the issue.

BUG=b:200743117,b:188454947
TEST=Run perf test against Nami.

Change-Id: I00a6a38ac22e8049063b424bad07d70b0b2e044c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3846103
Tested-by: Kevin Lund <kglund@google.com>
Reviewed-by: Arowa Suliman <arowa@chromium.org>
Commit-Queue: Kevin Lund <kglund@google.com>