audio_AudioAfterSuspend: verify DUT is connectable on resume
This makes audio_AudioAfterSuspend.headphone raise error instead of
failing when it fails to resume, since it's not caused by the audio.

We need verify_connectivity because is_up does not check if the DUT is
connectable (SSHTimeout), which is critical to running the subsequent
test.

BUG=b:234549189
TEST=test_that --... audio_AudioAfterSuspend.headphone

Change-Id: I6ed756b6df3ab1a006f9d04ed22b20734f35a9e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3852377
Commit-Queue: Jeremy Wu <jrwu@google.com>
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Tested-by: Jeremy Wu <jrwu@google.com>