Bluetooth: Add a new test to check qr states
This new test checks whether the Bluetooth Quality Report feature
is enabled or disabled as expected in various scenarios when
the DUT reboots, the Bluetooth daemon restarts, the adapter powers
cycle, and the system suspends and resumes.

This test checks the Bluetooth software stack and does not check
the controller functionality. Hence, a btpeer device is not needed
in this test.

BUG=b:243625074
TEST=Run the qr_check_states_test
(cr) test_that $DUT_IP
     bluetooth_AdapterQRHealth.qr_check_states_test

Change-Id: Ic7933b530557482ad6c9cc7d083cc24df7b439d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3855387
Reviewed-by: Dean Liao <deanliao@chromium.org>
Reviewed-by: Archie Pusaka <apusaka@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>