Temporarily maintain py2 compatibility for autotest client
autoupdate_EndToEndTest is still running on the drone, and
autotest-drone still has the code path to run the client in py2 for
milestones < M100. However some code introduced in M105 and onwards is
no longer py2 compatible. autotest still uses the latest client code
regardless of the DUT version that autoupdate_EndToEndTest provisions,
so we get in a bad scenario where we're running >M105 client code on
<M100 client versions with python 2.

As a quick fix to unblock M105 stable RCs, we can change a few sections
to make them keep working on python 2.

The SSP (non-drone) version of autoupdate_EndToEndTest is being
scheduled in the lab now (CL:3857491). Once we confirm that the test is
running fine with SSP enabled, we can completely switch over the test to
use SSP and then revert this CL (since client will no longer run in
py2).

BUG=b:244594800
TEST=autoupdate_EndToEndTest with autotest.py modified to use py2 for
<M100 (this modification: https://paste.googleplex.com/5160679295156224)
TEST=test_that chromeos6-row10-rack24-host7 autoupdate_EndToEndTest -b octopus --autotest_dir ~/chromiumos/src/third_party/autotest/files --args="update_type=delta source_release=14268.67.0 source_payload_uri=gs://chromeos-releases/stable-channel/octopus/14268.67.0/payloads/chromeos_14268.67.0_octopus_stable-channel_full_test.bin-gyywenzrmiztsfqgbf5aypgylns47y5r target_release=14989.80.0 target_payload_uri=gs://chromeos-throw-away-bucket/stable-channel/octopus/14989.80.0/payloads/chromeos_14268.67.0-14989.80.0_octopus_stable-channel_delta_test.bin-gyztcmbsga4wiokf4w5uw6h2zu64uhot"

Change-Id: I93938f253b5f66d1bec2f8bf027b3f1ff34ca20b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3869349
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>
Commit-Queue: Kyle Shimabukuro <kyleshima@chromium.org>