power_status: Add error handling to MeasurementLogger
MeasurementLogger is used to collect information periodically on DUT
during the test. Currently, there is no any error handling to the
refresh function, which will invoke the measurement through running
commands or reading files. If there are any exceptions during the
measurement, the logger will crash silently and stop recording future
status unless we dig into the debug outputs, making the final results
inaccurate.

For example, PowerLogger, which is a child class of MeasurementLogger,
will run power_supply_info to collect battery status and energe usage.
However, this command seldom fails due to some issues. As a result, the
value of loop00_system_pwr_avg and wh_energy_powerlogger would be the
average energy rate of first few parts of the test, which is often much
higher than the overall average.

We should add error handling to the refresh function in
MeasurementLogger to prevent the thread from crashing. Instead, it will
only ignore a single measurement on refresh failure, which should not
affect the overall results much.

BUG=b:239161339
TEST=test_that power_LoadTest
     observe a refresh error without crash and it retries successfully
     locally run some unit tests for the case of retrying failure

Change-Id: Ib8b72f1661d19c206a7967da15750bbebf5e2c1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3889956
Tested-by: Yu-Che Cheng <giver@chromium.org>
Commit-Queue: Yu-Che Cheng <giver@chromium.org>
Reviewed-by: Fei Shao <fshao@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>