Adding new suite:wifi_matfunc_openwrt
Added new suite:wifi_matfunc_openwrt with tast tests to run on the pool
wificell_openwrt. This pool will run all the matfunc tests in tasts but
the expectation is that the only wificell test fixtures executing this
test are those with openwrt routers and pcaps.

BUG=b:245990547
TEST=Presubmit checks pass.

Change-Id: I10c7e346b0827727e915b143a33d9e9a6ac0c4c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3892030
Reviewed-by: Dinesh Kumar Sunkara <dsunkara@google.com>
Reviewed-by: Harpreet Grewal <harpreet@chromium.org>
Commit-Queue: Billy Zhao <billyzhao@chromium.org>
Tested-by: Billy Zhao <billyzhao@chromium.org>