au_e2e: wait for UI to stabilize post-reboot in test and provisioning
If any FW updates run when the device boots, they will prevent
update-engine from starting until the FW update completes. This can
cause problems after the reboots in provisioning and in the test itself.

The provisioning case is handled on the client side in kernel_utils.py
for client versions >=M104. This CL also waits for the UI to stabilize
from the server side in provisioner.py before calling into kernel_utils,
since we need it in earlier milestones as well.

It also adds it to au_e2e before checking update-engine's status
post-reboot, since it's also needed there.

A FW update for the touchpad shows up in both cases for brya/redrix with
the versions in the TEST section below, so it can be used to reliably
reproduce and test the behavior.

BUG=b:245396689
TEST=test_that <dut> autoupdate_EndToEndTest --args="update_type=delta source_release=14526.84.2 source_payload_uri=gs://chromeos-releases/stable-channel/brya/14526.84.2/payloads/chromeos_14526.84.2_brya_stable-channel_full_test.bin-gyzdkzjtgjtddyji7k2g2mqi2ex2l4vv target_release=14909.132.0 target_payload_uri=gs://chromeos-releases/stable-channel/brya/14909.132.0/payloads/chromeos_14526.84.2-14909.132.0_brya_stable-channel_delta_test.bin-gyztaobwge2tcmsyaopgwmbfio56b3m7"

Change-Id: Ia239e67aa4def324e60708090859559f439a157d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3894933
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>
Commit-Queue: Kyle Shimabukuro <kyleshima@chromium.org>