audio_facade: Ignore wrong audio routing error
When a device has a DependentPCM in alsa UCM config, CRAS makes the
device and its dependent as the same device. It makes
get_selected_output_device from CRAS to be wrong. Hence, we should catch
the thrown wrong routing exception and print it on the logs instead.

BUG=b:243326941
TEST=test_that audio_AudioVolume.headphone

Change-Id: I9288077316a9182dafdd12a75dcc0ccd344bebfc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3897997
Tested-by: Norman Bintang <normanbt@chromium.org>
Reviewed-by: Judy Hsiao <judyhsiao@google.com>
Commit-Queue: Norman Bintang <normanbt@chromium.org>
Reviewed-by: Archie Pusaka <apusaka@chromium.org>