autotest: Don't wait for the update screen in non critical update tests
autoupdate_NonBlockingOOBEUpdate fails on some builds to wait for
the update screen to be visible. The update screen will only be visible
while on the loading step for a short period time before it gets
skipped if the update is non critical.

Restrict the waiting for the update screen to tests testing the
critical updates, and wait for the next screen (user creation)
for tests testing the non critical updates.

Bug: b:241104698
Change-Id: I82f9f7f977b041083fcdc15db18fa6021caf676b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3899471
Commit-Queue: Osama Fathy <osamafathy@google.com>
Tested-by: Osama Fathy <osamafathy@google.com>
Reviewed-by: Roman Sorokin <rsorokin@google.com>