cheets_xTS: Support setting expected test count for "collect" jobs.
The "collect-tests-only" mode is not well maintained in upstream
and occasionally contains crash errors. In the past we implemented
a workaround to ignore such errors in crrev.com/c/2480302.

The current CL is for adding the workaround via the control file
generator, instead of manual editing.

BUG=b:181176656
TEST=./generate_controlfiles_GTS.py --is_latest $url

Change-Id: I9d86f98778981b71b97fe95d9f8628d5be740bec
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3899591
Reviewed-by: Rintaro Kuroiwa <rkuroiwa@google.com>
Auto-Submit: Kazuhiro Inaba <kinaba@chromium.org>
Tested-by: Kazuhiro Inaba <kinaba@chromium.org>