power_status: fix battery_discharge_ok_on_ac:
The method battery_discharge_ok_on_ac incorrectly checked that the
battery was charging to return true, when it should instead check that
the system has AC plugged in. This, coupled with the check that charge
is 90% or higher, avoids the issue with battery learning causing
spurious errors with this test.

BUG=b:217390814
TEST=power_ProbeDriver.probe_ac on atlas

Change-Id: I0890cb0b00e8e63775f52ebf617fe10706832c52
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3910710
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Auto-Submit: Derek Basehore <dbasehore@chromium.org>