audio_AudioAfterSuspend: raise if suspend time was shorter than expected
Some tests are failing due to remote/host.py suspending less than
expected. That should have been raised as an error, but was difficult to
pass to the main thread due to the design pattern.

In this CL, we instead verify that it has suspended no less than the
expected time in the main thread.

BUG=b:234549189
TEST=test_that audio_AudioAfterSuspend.headphone

Change-Id: I8ddfddc6ec4225e07deb1bf56f141b1fa3fcd018
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3912192
Tested-by: Jeremy Wu <jrwu@google.com>
Reviewed-by: Li-Yu Yu <aaronyu@google.com>
Commit-Queue: Jeremy Wu <jrwu@google.com>