autotest_client_httpd: Add a handler for OPTIONS method
There are only handlers for GET, POST, and HEAD methods in FormHandler,
while there is no handler for OPTIONS method.
In PLT tests, the client may send 'OPTIONS /scroll' HTTP requests to
this httpd service due to CORS policy, and the HTTP server will respond
'501 Not Implemented' and print an error message in the logs.
Although the response does not affect the subsequent 'GET /scroll'
request, it would be better to handle this request and prevent from
lots of error messages about unsupported request method.
Since there is nothing to do with the OPTIONS method, we simply respond
'204 No Content' to the requests.

BUG=b:248460013
TEST=test_that power_LoadTest
     tcpdump -i lo

Change-Id: I261ec995cf0664b2274779d27066cdf9454dca0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3921047
Tested-by: Yu-Che Cheng <giver@chromium.org>
Commit-Queue: Yu-Che Cheng <giver@chromium.org>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>
Reviewed-by: Fei Shao <fshao@chromium.org>