power_LoadTest: correctly respond to each HTTP request
Currently, most of the HTTP request handler does not correctly send a
response back to the client. Although these requests are only used to
send some information to the server and the client does not need the
response, it makes the client need to perform lots of error handling and
generate lots of annoying error messages.
We should correctly reply to each HTTP request.

BUG=b:237634370
TEST=test_that power_LoadTest

Change-Id: I7c08868e90f76aaa305669fec96b55b8fa5ea0f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3925496
Commit-Queue: Yu-Che Cheng <giver@chromium.org>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>
Tested-by: Yu-Che Cheng <giver@chromium.org>
Reviewed-by: Fei Shao <fshao@chromium.org>