bluetooth_wifi_coex: Adjust drop level expectations
Based on the staineless results with over 1000 fails due to the
measured drop level higher than expected the adjustment was made.
There was either a generic update or pe- board exception.

To make this possible, a simple parser of the stainless results was
created: https://paste.googleplex.com/6327286314827776

For every test variant (test type, AP config, BT state) boards with
exceeding drop level were grouped to the number of levels they are
over the requirement.

For the generic update, there must be at least five boards on the
same level above the requirement. If less, an exception was added
for every board.

BUG=b:224781062
TEST=test_that --args "btpeer_host_list=${RASPBERRY_IP} \
            router_addr=${AP_IP} pcap_addr=${PCAP_IP}" ${DUT_IP} \
            suite:bluetooth_wifi_coex

Change-Id: Id455967dfdf796f6931479185345a5a2a374688c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3924091
Commit-Queue: Jacek Siuda <jsiuda@google.com>
Reviewed-by: Arowa Suliman <arowa@chromium.org>
Tested-by: Jacek Klimkowicz <jakl@semihalf.com>