power_LoadTest: check runtime.lastError in API callback
The PLT extension communicates with each tab through sendMessage API
during the test. However, there is some race condition that it may try to
send a message to a tab which has not successfully loaded and registered
the event listener yet. In this case, an error will be set at
chrome.runtime.lastError, and the browser expects the API callback to
check for the error. If an error occurs but hasn't been checked (evaluated)
in the callback function, the error will be raised to the top level. In
our case, an error message will be sent and logged in the debug logs.
We should check the error in its API callback to prevent from the
logs of error logs even if the error is expected and we don't need to
use it at all.

BUG=b:237634370
TEST=test_that power_LoadTest

Change-Id: Ia8df7962074f20e6b14e9469047d70b88e1031e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3928450
Reviewed-by: Fei Shao <fshao@chromium.org>
Tested-by: Yu-Che Cheng <giver@chromium.org>
Commit-Queue: Yu-Che Cheng <giver@chromium.org>