autotest: Use the correct interface for BT audio streaming tests
When DUT is connected to Satlab and tests are started from local network
with ssh tunneling, a real DUT IP is required. As the DUT interface
enumaration is non-deterministic during boot time so it may vary.
Another interface is present when i.e. servo is connected. The active
interface is picked up instead of hardcoded one.

BUG=b:251378000
TEST=test_that --args "btpeer_host_list=${RASPBERRY_IP} \
        router_addr=${AP_IP} pcap_addr=${PCAP_IP}" ${DUT_IP} \
        suite:bluetooth_wifi_coex

Change-Id: I792b98b1c9bb86c4c2a4f259a2d17f4e2c979495
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3938067
Tested-by: Jacek Klimkowicz <jakl@semihalf.com>
Reviewed-by: Kalin Stoyanov <kalin@chromium.org>
Commit-Queue: Jacek Siuda <jsiuda@google.com>
Reviewed-by: Dean Liao <deanliao@chromium.org>