bluetooth: fix qr_disabled_a2dp_test
In the qr_disabled_a2dp_test, there is no need to enable the
quality report and the quality debug log in qr_disabled_a2dp()
since they must be enabled in run_test_method() before pairing
is performed.

On the other hand, we need to disable the quality report and
keep the quality debug log enabled for the test to verify that
no BQR events are received.

BUG=b:252864798
TEST=Run the tests.

[Single test]: bluetooth_AdapterQRHealth.qr_disabled_a2dp_test
  - The BQR events were observed when Raspberry Pi was connected.
  - When running qr_disabled_a2dp(), no more BQR events were observed.

[All QR tests]: all bluetooth_AdapterQRHealth tests passed.

Change-Id: Ib0b6631be29a7b447ca2a1a4c2ca880cb28132b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3955103
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Archie Pusaka <apusaka@chromium.org>
Reviewed-by: John Lai <johnlai@google.com>