Fix test because expect behavior change
The test originally expects each time the volume decrease/increase by 4, but the volume is no longer a fixed decrease/increase 4, but a multiple of 4 to the most recent volume.

In this example the test starts at 75 and then drops to 72 instead of 71.

Improve this test so that it doesn't rely on confirming whether the volume goes down or up by a fixed value

BUG=b:252569317
TEST=pass platform_InputVolume

Change-Id: Id6ae3b19683ccf4e90c18857cc91b96f1ce460f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3954189
Commit-Queue: Ching Yun Chang <whalechang@google.com>
Reviewed-by: En-Shuo Hsu <enshuo@chromium.org>
Tested-by: Ching Yun Chang <whalechang@google.com>