[Tauto] Make TIME an optional value
Use 'default' for any tests which do not declare a TIME.  Eventually
we should remove TIME entirely, but it would be easier to do that
after in-planning test scheduling changes are made.  In the meantime,
make it optional so tests can ignore that the value exists.

BUG=None
TEST=ran generate_metadata.py

Change-Id: Ic2d4d111ec1607f418cc0c1002afdacdc1f63102
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3939655
Reviewed-by: Derek Beckett <dbeckett@chromium.org>
Tested-by: Katherine Threlkeld <kathrelkeld@chromium.org>
Commit-Queue: Katherine Threlkeld <kathrelkeld@chromium.org>