power_status: Fix GPUFreqStats failing if CAGF is absent on i915.
On i915 platforms, GPUFreqStats assumes the current frequency can always
be found in a "CAGF" line in /run/debugfs_gpu/i915_frequency_info. On
5.15 with i915.enable_guc=3, this line is absent, causing power tests to
fail. Match on "Actual freq" as well (given it matches the CAGF value
when both are present), and take the first of the two found. CAGF occurs
earlier in the file so will take priority, leaving the detected current
frequency unchanged when it is present.

BUG=b:252928561
TEST=power_Speedometer2 on brya and nissa logs reasonable current GPU
     frequency values and GPU frequency buckets

Change-Id: Ie49593852fd12e76c00bddbadc26c2bf80436a7d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3971826
Reviewed-by: Victor Ding <victording@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Commit-Queue: Sam McNally <sammc@chromium.org>
Tested-by: Sam McNally <sammc@chromium.org>