bluetooth: polling bluetoothd in qr_enable_quality_report
After restarting bluetoothd, it is required to poll and check if
bluetoothd is ready. Otherwise, the following dbus method, e.g.,
self.enable_disable_quality_debug_log(enable=True), may fail.

This patch also groups a few common BQR enabling/disabling operations
into the sequences, i.e., qr_enable_quality_report() and
qr_disable_quality_report(), so that they can be called consistently
in bluetooth_AdapterQRHealth and bluetooth_AdapterSRHealth.

It is also important to place qr_disable_quality_report() in the
"finally" clause of a try block. This is to ensure that the BQR can
be disabled even when an exception occurs in the test method.

BUG=b:255666931
TEST=The following tests passed.
- all tests in bluetooth_AdapterQRHealth including qr_check_states_test
- all tests in bluetooth_AdapterSRHealth

Change-Id: I75d061e144c8e74ff15c8759ee7acc78f12f53fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3985183
Reviewed-by: Archie Pusaka <apusaka@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>