power_WakeSources: increase reconnect timeout
Due to lab ethernet dongles potentially restarting in storage mode, we
need to increase the timeout for waiting for the DUT to be up to 3
minutes. This should only impact test time when things are going wrong,
but that may be happening.

BUG=b:255880142
TEST=not tested

Change-Id: Ibe07ffd9c3e9c93e062712455d3db3c11f1775c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3985326
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>
Auto-Submit: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>