Add CPU balance and frequency of high consumption
Check if any CPUs do not have any Softirqs and display
a message that the CPUs are not balanced if there are CPUs
without any Softirqs. Track the frequency with which each
CPU exceeds the CPU consumption threshold over the polling
period.

BUG=b:254326604, b:257075290
TEST=Tested by calling network_WiFi_Perf and verifying the
logging output.

Change-Id: Icb7f48e8565559d2bc891480390206ebe4620894
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3991746
Commit-Queue: Andrew DeBruin <adebruin@chromium.org>
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>
Reviewed-by: Billy Zhao <billyzhao@chromium.org>
Tested-by: Andrew DeBruin <adebruin@chromium.org>