Lacros version skew: Add Ash and Lacros versions to reporting output.
As Lacros is provisioned inside of the autotest file, we want to expose
Lacros version (together with ash versions) to keyval which ultimately
will be written to RDB. This allows us to create reports to visualize
the test results of version skew (Lacros against ash version).

Version for a channel is typically looked up from the VersionHistory APIs
For the case where user supplies gcs path for lacros artifacts directly,
we will find out the version from the metadata.json in the unzipped
contents.

BUG=b:236018330
TEST=test_that --board={BOARD} --args="tast_expr='lacros.Basic'" {DUT} tast.lacros-version-skew-canary
TEST=test_that --board={BOARD} --args="tast_expr='lacros.Basic' lacros_gcs_path='gs://chrome-unsigned/desktop-5c0tCh/106.0.5246.0/lacros64/lacros.zip'" {DUT} tast.lacros-version-skew-gspath

Change-Id: I46dadffadca678898a6a6016e5e4c28235d41517
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3989573
Reviewed-by: Derek Beckett <dbeckett@chromium.org>
Reviewed-by: Hyung Tae Kim <hyungtaekim@chromium.org>
Commit-Queue: Jonathan Fan <jonfan@chromium.org>
Tested-by: Jonathan Fan <jonfan@chromium.org>