Reset abi information to avoid an error in GTS testing
In GTS, _ABI_LIST is not in the key of bundle_url_config.json, but abi may be specified from the control files.
This raises an error to make a bundle url in bundle_utils.py
This CL deletes an error handling to avoid errors in GTS testing, and reset abi to handle it correctly.

BUG=b:256079546
TEST=test_that -b $BOARD localhost:2222 cheets_GTS_R.internal.arm.GtsMediaTestCases.32 after reverting commit chrome-internal:5061569
TEST=./utils/unittest_suite.py server.cros.tradefed.bundle_utils_unittest

Change-Id: I18f5d10ea1017f4ea3e916fe098ae3a818d0f574
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3992950
Commit-Queue: Shuhei Kobayashi <shukobayashi@google.com>
Reviewed-by: Rintaro Kuroiwa <rkuroiwa@google.com>
Tested-by: Shuhei Kobayashi <shukobayashi@google.com>