sys_power: Allow cr50-related i2c transfer warnings during suspend.
According to b/250089565#comment6 and related bugs, this warning is not
feasible to fix in the short term and is relatively innocuous.

In this case the log line that indicates the warning is cr50-related can
be much later, so scan ahead to the end of the warning stack trace.

BUG=b:250089565, b:251736428, b:223442909, b:250603092, b:234483910
TEST=power_SuspendStress.bareFSI on various nissa models

Change-Id: Ic3f42a4c9b60282a8e88699959acad5c86d84eb7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4011080
Tested-by: Sven van Ashbrook <svenva@google.com>
Tested-by: Sam McNally <sammc@chromium.org>
Reviewed-by: Sven van Ashbrook <svenva@google.com>
Commit-Queue: Sven van Ashbrook <svenva@google.com>
Reviewed-by: Lalith Rajendran <lalithkraj@google.com>
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>