Tool for upreving xTS
When the latest xTS is released, the version number needs to be edited, and the test needs to be uploaded to gs.
This CL automates this process with a single command.
The commands are as follows:
- python uprev_official.py --to 11_r9
- python uprev_preview.py --to 9259352
- python uprev_preview.py --to_latest
This CL adds the function to download the specified test suites to a temporary directory, modify the corresponding build id in bundle_url_config.json and upload it to gs if necessary.

BUG=b:181196484
TEST=python uprev_preview.py --to_latest
TEST=python uprev_preview.py --to 9259352
TEST=./utils/unittest_suite.py server.cros.tradefed.bundle_utils_unittest
TEST=./utils/unittest_suite.py server.cros.tradefed.uprev_preview_version_common_unittest

Cq-Depend: chrome-internal:5047351
Change-Id: I8545c5107c35cc5a480b4f824b10e55842360d53
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3966289
Reviewed-by: Kazuhiro Inaba <kinaba@chromium.org>
Tested-by: Shuhei Kobayashi <shukobayashi@google.com>
Commit-Queue: Shuhei Kobayashi <shukobayashi@google.com>