power_VideoCall: Update option file to resolve argument bug
disable_hdrnet was added to the control.option file but that variable
was never declared, causing a runtime error when trying to run the
test.  Add the missing declaration and organize the file to make
it clear where variables are declared, and where they are over
ridden.

BUG=None
TEST=test_that dut_ip power_VideoCall.option

Change-Id: I55b616eb5be73d38ef28098abe1fb64bcfa3a41b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4024072
Commit-Queue: Jonathon Murphy <jpmurphy@google.com>
Auto-Submit: Jonathon Murphy <jpmurphy@google.com>
Reviewed-by: Tim Van Patten <timvp@google.com>
Tested-by: Jonathon Murphy <jpmurphy@google.com>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>