Bluetooth: Disable reboot standalone test as persistence is flaky
Currently, this test is not supported for floss because we use the
chrome flag to switch to the floss. After the reboot, Chrome will
use Bluez directly and might not switch to the floss. We can re-enable
it after we can make the floss to be the default Bluetooth daemon.

BUG=b:173248527
TEST=None

Change-Id: I8573824ec13e6f0ea69c2e63c07098e452df269f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4032469
Commit-Queue: John Lai <johnlai@google.com>
Tested-by: John Lai <johnlai@google.com>
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>