power_WakeSources: don't disable sysinfo collection
Switch this test to the default behaviour (i.e. collect sysinfo after
the test) so that we can see powerd logs, syslog, and kernel messages
in order to diagnose failures.

Although the sysinfo collection is somewhat slow and heavyweight
(~1 minute) it's dwarfed by the time taken for the test itself (several
minutes) and we desperately need more information about the failure
scenarios for this test so that we can improve its pass rate. So the
tradeoff seems worthwhile.

BUG=b:259630376
TEST=test_that $DUT power_WakeSources

Change-Id: I881ee5807dbf99b2af4d6bf4d2c652173fe95a8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4041682
Tested-by: Dan Callaghan <dcallagh@chromium.org>
Reviewed-by: Derek Basehore <dbasehore@chromium.org>
Commit-Queue: Dan Callaghan <dcallagh@chromium.org>