FAFT: Skip firmware_CsmeFwUpdate only if me_rw is not found
The me_rw blob only exists on Intel platforms. Previously the test
checked if faft_config.ec_capability contained x86 to decide whether
to skip the test. However, that was problematic because the test
should also be skipped on AMD platforms.

CL:3954709 tried to fix the problem with get_intel_cpu_uarch(), but
that function actually returns the architecture of the autotest host,
not the client.

Now we want to try a different solution. Since we are already skipping
the test if me_rw is not found in the firmware image, the extra check
of the client CPU architecture seems unnecessary. An experimental run
on frostflow shows that removing the check increases only 4.1s of the
test running time (for dumping the firmware and checking the CBFS),
which is insignificant compared to the total time.

Therefore, remove the incorrect get_intel_cpu_uarch check and rely on
check_if_me_blob_exist_in_image to do its work. Slightly adjust the code
order to minimize the time wasted on non-Intel platforms.

BUG=b:253552317, b:259351173
TEST=verify on Intel and AMD platform

Change-Id: I32e03e2614a93d039e982af16683c9f0957c2ea8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/4048544
Tested-by: John Su <john_su@compal.corp-partner.google.com>
Reviewed-by: Chao Gui <chaogui@google.com>
Auto-Submit: John Su <john_su@compal.corp-partner.google.com>
Commit-Queue: Chao Gui <chaogui@google.com>
Reviewed-by: Karthikeyan Ramasubramanian <kramasub@google.com>
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>